Move pushdown test to `BaseRedisConnectorTest`
